Refactor helpers into a common folder
